Use writable URL when cloning map repos
The Git protocol does not support authentication, and thus cannot be used
to push changes back to GitHub.  The clone_all.sh script was modified to
clone all repos using the SSH protocol, which does permit changes to be
pushed back to GitHub (assuming the user has registered their SSH key
with GitHub).